refactor(menu, menuGroup) : 메뉴, 메뉴그룹 패키지 분리
