+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,6 +1,9 @@
 package kitchenpos;
 
 import kitchenpos.domain.*;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
